refactor(geometry): split processors/brep.rs into faceted + surface_model
brep.rs (833 LOC) -> brep/{faceted.rs (486: FacetedBrepProcessor + shared
triangulate_face raised to pub(super)), surface_model.rs (352: the two
near-duplicate ConnectedFaceSet/Shell processors), mod.rs (re-exports)}.
Pure code motion; processors/mod.rs untouched, all paths re-exported.
faceted.rs stays >400 as one cohesive processor+triangulation unit (no inline
tests to relocate). build/clippy -D warnings clean; brep tests green.
